Refactored all seriaizers into their own file
