+++ infraohjelmointi_api/serializers/BaseMeta.py
@@ -1,0 +1,2 @@
+class BaseMeta:
+    exclude = ["createdDate", "updatedDate"]
