+++ infraohjelmointi_api/serializers/NotePersonSerializer.py
@@ -1,0 +1,8 @@
+from infraohjelmointi_api.models import Person
+from rest_framework import serializers
+
+
+class NotePersonSerializer(serializers.ModelSerializer):
+    class Meta:
+        fields = ("id", "firstName", "lastName")
+        model = Person
