+++ infraohjelmointi_api/serializers/ProjectGetSerializer.py
@@ -1,0 +1,81 @@
+from infraohjelmointi_api.models import Project
+from infraohjelmointi_api.serializers import (
+    BaseMeta,
+    DynamicFieldsModelSerializer,
+    PersonSerializer,
+    ProjectLockSerializer,
+)
+from infraohjelmointi_api.serializers.BudgetItemSerializer import BudgetItemSerializer
+from infraohjelmointi_api.serializers.ConstructionPhaseDetailSerializer import (
+    ConstructionPhaseDetailSerializer,
+)
+from infraohjelmointi_api.serializers.ConstructionPhaseSerializer import (
+    ConstructionPhaseSerializer,
+)
+from infraohjelmointi_api.serializers.PlanningPhaseSerializer import (
+    PlanningPhaseSerializer,
+)
+from infraohjelmointi_api.serializers.ProjectAreaSerializer import ProjectAreaSerializer
+from infraohjelmointi_api.serializers.ProjectCategorySerializer import (
+    ProjectCategorySerializer,
+)
+from infraohjelmointi_api.serializers.ProjectPhaseSerializer import (
+    ProjectPhaseSerializer,
+)
+from infraohjelmointi_api.serializers.ProjectPrioritySerializer import (
+    ProjectPrioritySerializer,
+)
+from infraohjelmointi_api.serializers.ProjectQualityLevelSerializer import (
+    ProjectQualityLevelSerializer,
+)
+from infraohjelmointi_api.serializers.ProjectResponsibleZoneSerializer import (
+    ProjectResponsibleZoneSerializer,
+)
+from infraohjelmointi_api.serializers.ProjectRiskSerializer import ProjectRiskSerializer
+from infraohjelmointi_api.serializers.ProjectSetCreateSerializer import (
+    ProjectSetCreateSerializer,
+)
+from infraohjelmointi_api.serializers.ProjectTypeSerializer import ProjectTypeSerializer
+from rest_framework import serializers
+
+
+class ProjectGetSerializer(DynamicFieldsModelSerializer):
+    projectReadiness = serializers.SerializerMethodField()
+    projectSet = ProjectSetCreateSerializer(read_only=True)
+    siteId = BudgetItemSerializer(read_only=True)
+    area = ProjectAreaSerializer(read_only=True)
+    type = ProjectTypeSerializer(read_only=True)
+    priority = ProjectPrioritySerializer(read_only=True)
+    phase = ProjectPhaseSerializer(read_only=True)
+    personPlanning = PersonSerializer(read_only=True)
+    personProgramming = PersonSerializer(read_only=True)
+    personConstruction = PersonSerializer(read_only=True)
+    estPlanningStart = serializers.DateField(format="%d.%m.%Y")
+    estPlanningEnd = serializers.DateField(format="%d.%m.%Y")
+    category = ProjectCategorySerializer(read_only=True)
+    constructionPhaseDetail = ConstructionPhaseDetailSerializer(read_only=True)
+    riskAssessment = ProjectRiskSerializer(read_only=True)
+    estConstructionStart = serializers.DateField(format="%d.%m.%Y")
+    estConstructionEnd = serializers.DateField(format="%d.%m.%Y")
+    presenceStart = serializers.DateField(format="%d.%m.%Y")
+    presenceEnd = serializers.DateField(format="%d.%m.%Y")
+    visibilityStart = serializers.DateField(format="%d.%m.%Y")
+    visibilityEnd = serializers.DateField(format="%d.%m.%Y")
+    constructionPhase = ConstructionPhaseSerializer(read_only=True)
+    planningPhase = PlanningPhaseSerializer(read_only=True)
+    projectQualityLevel = ProjectQualityLevelSerializer(read_only=True)
+    responsibleZone = ProjectResponsibleZoneSerializer(read_only=True)
+    locked = serializers.SerializerMethodField()
+
+    class Meta(BaseMeta):
+        model = Project
+
+    def get_locked(self, obj):
+        try:
+            lockData = ProjectLockSerializer(obj.lock, many=False).data
+            return lockData
+        except:
+            return None
+
+    def get_projectReadiness(self, obj):
+        return obj.projectReadiness()
