+++ infraohjelmointi_api/serializers/ProjectLockSerializer.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from ..models import ProjectLock
+from .BaseMeta import BaseMeta
+
+
+class ProjectLockSerializer(serializers.ModelSerializer):
+    class Meta(BaseMeta):
+        model = ProjectLock
