+++ infraohjelmointi_api/serializers/ProjectSetGetSerializer.py
@@ -1,0 +1,35 @@
+from infraohjelmointi_api.models import ProjectSet
+from infraohjelmointi_api.serializers import BaseMeta
+from infraohjelmointi_api.serializers.ProjectPhaseSerializer import (
+    ProjectPhaseSerializer,
+)
+from rest_framework import serializers
+
+from django.db.models import Q
+
+
+class ProjectSetGetSerializer(serializers.ModelSerializer):
+    sapProjects = serializers.SerializerMethodField()
+    sapNetworks = serializers.SerializerMethodField()
+    phase = ProjectPhaseSerializer(read_only=True)
+
+    class Meta(BaseMeta):
+        model = ProjectSet
+
+    def get_sapNetworks(self, obj):
+        return [
+            network
+            for obj in obj.project_set.all()
+            .filter(~Q(sapNetwork=None))
+            .values("sapNetwork")
+            for network in obj["sapNetwork"]
+        ]
+
+    def get_sapProjects(self, obj):
+        return [
+            project
+            for obj in obj.project_set.all()
+            .filter(~Q(sapNetwork=None))
+            .values("sapProject")
+            for project in obj["sapProject"]
+        ]
